Revamped the map linking
Every element in the map is completely independent, except for the
targets.

Every map element has a function 'target' which returns a function
which evaluates a cars next target. Due to the cyclic nature of these
targets, they are linked in every map element after initialization as
the 'nt' variable.

Additionally, Kevin pointed out that Cars will also need access to the
map, so now every map element is linked to the 'map' variable when
added to the map. The maps 'add' and 'remove' methods can be used to
handle this.